# Covenant/API/Models/CodeMirrorTheme.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace EasyPeasy.API.Models
{
    using Newtonsoft.Json;
    using Newtonsoft.Json.Converters;
    using System.Runtime;
    using System.Runtime.Serialization;

    /// <summary>
    /// Defines values for CodeMirrorTheme.
    /// </summary>
    [JsonConverter(typeof(StringEnumConverter))]
    public enum CodeMirrorTheme
    {
        [EnumMember(Value = "default")]
        Default,
        [EnumMember(Value = "night")]
        Night
    }
    internal static class CodeMirrorThemeEnumExtension
    {
        internal static string ToSerializedValue(this CodeMirrorTheme? value)
        {
            return value == null ? null : ((CodeMirrorTheme)value).ToSerializedValue();
        }

        internal static string ToSerializedValue(this CodeMirrorTheme value)
        {
            switch( value )
            {
                case CodeMirrorTheme.Default:
                    return "default";
                case CodeMirrorTheme.Night:
                    return "night";
            }
            return null;
        }

        internal static CodeMirrorTheme? ParseCodeMirrorTheme(this string value)
        {
            switch( value )
            {
                case "default":
                    return CodeMirrorTheme.Default;
                case "night":
                    return CodeMirrorTheme.Night;
            }
            return null;
        }
    }
}