# Covenant/API/Models/StringIdentityUserRole.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace EasyPeasy.API.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class StringIdentityUserRole
    {
        /// <summary>
        /// Initializes a new instance of the StringIdentityUserRole class.
        /// </summary>
        public StringIdentityUserRole()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the StringIdentityUserRole class.
        /// </summary>
        public StringIdentityUserRole(string userId = default(string), string roleId = default(string))
        {
            UserId = userId;
            RoleId = roleId;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "userId")]
        public string UserId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "roleId")]
        public string RoleId { get; set; }

    }
}