# Covenant/API/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "EasyPeasy API",
    "version": "v0.1"
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "GetUsers",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/EasyPeasyUser"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "CreateUser",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EasyPeasyUserRegister"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/EasyPeasyUser"
            }
          }
        }
      },
      "put": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "EditUser",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EasyPeasyUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/EasyPeasyUser"
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "GetUser",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/EasyPeasyUser"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "DeleteUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/users/login": {
      "post": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "Login",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EasyPeasyUserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/EasyPeasyUserLoginResult"
            }
          }
        }
      }
    },
    "/api/users/logon": {
      "put": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "EditUserPassword",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EasyPeasyUserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/EasyPeasyUser"
            }
          }
        }
      }
    },
...
    "/api/users/{id}/roles": {
      "get": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "GetUserRoles",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/StringIdentityUserRole"
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/roles/{rid}": {
      "get": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "GetUserRole",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "rid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StringIdentityUserRole"
            }
          }
        }
      },
      "post": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "CreateUserRole",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "rid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StringIdentityUserRole"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "DeleteUserRole",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "rid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/roles/{rid}": {
      "get": {
        "tags": [
          "EasyPeasyUserApi"
        ],
        "operationId": "GetRole",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "rid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/IdentityRole"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CodeMirrorTheme": {
      "enum": [
        "default",
        "night"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "CodeMirrorTheme",
        "modelAsString": false
      }
    },
    "Theme": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "backgroundColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "backgroundTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "primaryColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "primaryTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "primaryHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "secondaryColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "secondaryTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "secondaryHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalBorderColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "navbarColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "sidebarColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputDisabledColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "textLinksColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "codeMirrorTheme": {
          "$ref": "#/definitions/CodeMirrorTheme"
        }
      }
    },
    "EasyPeasyUser": {
      "type": "object",
      "properties": {
        "themeId": {
          "format": "int32",
          "type": "integer"
        },
        "theme": {
          "$ref": "#/definitions/Theme"
        },
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "normalizedUserName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "normalizedEmail": {
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "passwordHash": {
          "type": "string"
        },
        "securityStamp": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "phoneNumberConfirmed": {
          "type": "boolean"
        },
        "twoFactorEnabled": {
          "type": "boolean"
        },
        "lockoutEnd": {
          "format": "date-time",
          "type": "string"
        },
        "lockoutEnabled": {
          "type": "boolean"
        },
        "accessFailedCount": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "EasyPeasyUserRegister": {
      "required": [
        "confirmPassword",
        "password",
        "userName"
      ],
      "type": "object",
      "properties": {
        "confirmPassword": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "EasyPeasyUserLogin": {
      "required": [
        "password",
        "userName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "EasyPeasyUserLoginResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "covenantToken": {
          "type": "string"
        }
      }
    },
    "StringIdentityUserRole": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "roleId": {
          "type": "string"
        }
      }
    },
    "IdentityRole": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "normalizedName": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}